modulizing canvas
