+++ frontend/src/app/modules/canvasEditor/canvas/CanvasBox.tsx
@@ -1,91 +1,24 @@
-import React, { useEffect, useRef } from "react";
 import { BoxProps } from "./BoxProps";
-import { createBoxRenderer } from "./BoxRenderer";
+import { useGlobalRefs } from "./hooks/useRef";
+import { useDraggable } from "./hooks/useEffect";
 
-export default function CanvasBox({ element, openInterface, updatePosition }: BoxProps) {
-  const gRef = useRef<SVGGElement>(null);
-  const isDragging = useRef(false);
-  const start = useRef({ x: 0, y: 0 });
-  const origin = useRef({ x: 0, y: 0 });
-  const halfSize = useRef({ w: 0, h: 0 });
+export default function CanvasBox({
+  element,
+  openInterface,
+  updatePosition,
+}: BoxProps) {
+  const { gRef, isDragging, start, origin, halfSize } = useGlobalRefs();
 
-  useEffect(() => {
-    if (!gRef.current) return;
-
-    const svgElement = createBoxRenderer(element);
-    const padding = 12;
-    gRef.current.innerHTML = "";
-    gRef.current.appendChild(svgElement);
-
-    const bbox = svgElement.getBBox();
-    const width = bbox.width + padding * 2;
-    const height = bbox.height + padding * 2;
-
-    svgElement.setAttribute("viewBox", `-${padding} -${padding} ${width} ${height}`);
-    svgElement.setAttribute("width", `${width}`);
-    svgElement.setAttribute("height", `${height}`);
-
-    halfSize.current = { w: width / 2, h: height / 2 };
-    gRef.current.setAttribute("transform", `translate(${element.x - halfSize.current.w}, ${element.y - halfSize.current.h})`);
-
-    const overlay = document.createElementNS("http://www.w3.org/2000/svg", "rect");
-    overlay.setAttribute("x", `-${padding}`);
-    overlay.setAttribute("y", `-${padding}`);
-    overlay.setAttribute("width", `${width}`);
-    overlay.setAttribute("height", `${height}`);
-    overlay.setAttribute("fill", "transparent");
-    overlay.style.cursor = "grab";
-
-    overlay.addEventListener("mousedown", onMouseDown as any);
-    overlay.addEventListener("click", (e) => {
-      e.stopPropagation();
-      openInterface(element);
-    });
-
-    svgElement.appendChild(overlay);
-  }, [element]);
-
-  const getSvgPoint = (e: MouseEvent | React.MouseEvent) => {
-    const svg = gRef.current!.ownerSVGElement!;
-    const pt = svg.createSVGPoint();
-    pt.x = e.clientX;
-    pt.y = e.clientY;
-    return pt.matrixTransform(svg.getScreenCTM()!.inverse());
-  };
-
-  const onMouseDown = (e: MouseEvent | React.MouseEvent) => {
-    e.stopPropagation();
-    isDragging.current = true;
-    const pt = getSvgPoint(e);
-    start.current = { x: pt.x, y: pt.y };
-    origin.current = { x: element.x, y: element.y };
-    window.addEventListener("mousemove", onMouseMove);
-    window.addEventListener("mouseup", onMouseUp);
-  };
-
-  const onMouseMove = (e: MouseEvent) => {
-    if (!isDragging.current) return;
-    const pt = getSvgPoint(e);
-    const dx = pt.x - start.current.x;
-    const dy = pt.y - start.current.y;
-
-    const svg = gRef.current!.ownerSVGElement!;
-    const vb = svg.viewBox.baseVal;
-    const { w, h } = halfSize.current;
-
-    const newX = clamp(origin.current.x + dx, vb.x + w, vb.x + vb.width - w);
-    const newY = clamp(origin.current.y + dy, vb.y + h, vb.y + vb.height - h);
-    updatePosition(newX, newY);
-  };
-
-  const onMouseUp = () => {
-    isDragging.current = false;
-    window.removeEventListener("mousemove", onMouseMove);
-    window.removeEventListener("mouseup", onMouseUp);
-  };
-
-  const clamp = (val: number, min: number, max: number) =>
-    Math.min(Math.max(val, min), max);
+  useDraggable(
+    gRef,
+    element,
+    halfSize,
+    openInterface,
+    isDragging,
+    start,
+    origin,
+    updatePosition
+  );
 
   return <g ref={gRef} />;
 }
